refac: 거래코드분리 && path 분리
