+++ server/swagger/swagger-output.json
@@ -11,7 +11,7 @@
       "url": "http://localhost:5000"
     }
   ],
-  "basePath": "/",
+  "basePath": "/api",
   "schemes": [
     "http"
   ],
@@ -28,7 +28,7 @@
     }
   ],
   "paths": {
-    "/buy": {
+    "/trade/buy": {
       "post": {
         "tags": ["구매"],
         "description": "구매자 정보 등록",
@@ -74,7 +74,7 @@
         }
       }
     },
-    "/SellerMatch": {
+    "/trade/SellerMatch": {
       "get": {
         "tags": ["거래"],
         "description": "추천 판매자",
@@ -91,7 +91,7 @@
         }
       }
     },
-    "/SellerMatch/{name}": {
+    "/trade/SellerMatch/{name}": {
       "post": {
         "tags": ["거래"],
         "description": "추천 판매자",
@@ -487,24 +487,7 @@
         }
       }
     },
-    "/user/updateAddress": {
-      "patch": {
-        "tags": ["사용자"],
-        "description": "사용자 주소 업데이트",
-        "responses": {
-          "200": {
-            "description": "OK"
-          },
-          "404": {
-            "description": "Not Found"
-          },
-          "500": {
-            "description": "Internal Server Error"
-          }
-        }
-      }
-    },
-    "/user/address": {
+    "/user/changeAddress": {
       "put": {
         "summary": "사용자 주소 정보 업데이트",
         "tags": ["User"],
@@ -614,6 +597,69 @@
                       "type": "string",
                       "example": "주소 업데이트 도중 에러가 발생하였습니다."
                     }
+                  }
+                }
+              }
+            }
+          }
+        }
+      }
+    },
+    "/mySells": {
+      "get": {
+        "summary": "내 판매 목록 조회",
+        "tags": ["판매"],
+        "security": [
+          {
+            "bearerAuth": []
+          }
+        ],
+        "responses": {
+          "200": {
+            "description": "판매 목록 조회 성공",
+            "content": {
+              "application/json": {
+                "schema": {
+                  "type": "array",
+                  "items": {
+                    "type": "object",
+                    "properties": {
+                      "id": { "type": "string", "example": "60d21b4667d0d8992e610c85" },
+                      "sellerId": { "type": "string", "example": "user123" },
+                      "title": { "type": "string", "example": "MacBook Pro 2021" },
+                      "price": { "type": "number", "example": 1500000 },
+                      "description": { "type": "string", "example": "사용감 적고 깨끗한 상태입니다." },
+                      "images": {
+                        "type": "array",
+                        "items": { "type": "string", "example": "data:image/png;base64,iVBORw0KGgo..." }
+                      }
+                    }
+                  }
+                }
+              }
+            }
+          },
+          "404": {
+            "description": "판매 중인 상품 없음",
+            "content": {
+              "application/json": {
+                "schema": {
+                  "type": "object",
+                  "properties": {
+                    "message": { "type": "string", "example": "판매중인 상품이 없습니다." }
+                  }
+                }
+              }
+            }
+          },
+          "500": {
+            "description": "서버 오류",
+            "content": {
+              "application/json": {
+                "schema": {
+                  "type": "object",
+                  "properties": {
+                    "message": { "type": "string", "example": "내 판매 목록을 불러오는 도중 에러가 발생하였습니다." }
                   }
                 }
               }
